+++ tests/application/interactors/__init__.py
@@ -1,0 +1,1 @@
+"""interactors."""
