+++ tests/application/interactors/test_fetch_task.py
@@ -9,9 +9,9 @@ import mock
 import pytest
 import requests
 
+from atcoder_helper.application.interactors.fetch_task import FetchTaskInteractor
 from atcoder_helper.application.usecases.errors import AtcoderAccessError
 from atcoder_helper.application.usecases.errors import ConfigAccessError
-from atcoder_helper.application.usecases.fetch_task import FetchTaskInteractor
 from atcoder_helper.entities.atcoder_task_config import TaskConfig
 from atcoder_helper.entities.atcoder_test_case import AtcoderTestCase
 from atcoder_helper.infrastructure.errors import ConnectionError
