--- tests/application/usecases/__init__.py
@@ -1,1 +1,0 @@
-"""usecases."""
